+++ controller/konnect/ops/ops_controlplane_test.go
@@ -8,9 +8,12 @@ import (
 	sdkkonnectcomp "github.com/Kong/sdk-konnect-go/models/components"
 	sdkkonnectops "github.com/Kong/sdk-konnect-go/models/operations"
 	sdkkonnecterrs "github.com/Kong/sdk-konnect-go/models/sdkerrors"
+	"github.com/google/uuid"
+	"github.com/samber/lo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8stypes "k8s.io/apimachinery/pkg/types"
 
 	"github.com/kong/gateway-operator/controller/konnect/conditions"
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
@@ -38,9 +41,11 @@ func TestCreateControlPlane(t *testing.T) {
 					},
 				}
 
+				expectedRequest := cp.Spec.CreateControlPlaneRequest
+				expectedRequest.Labels = WithKubernetesMetadataLabels(cp, expectedRequest.Labels)
 				sdk.
 					EXPECT().
-					CreateControlPlane(ctx, cp.Spec.CreateControlPlaneRequest).
+					CreateControlPlane(ctx, expectedRequest).
 					Return(
 						&sdkkonnectops.CreateControlPlaneResponse{
 							ControlPlane: &sdkkonnectcomp.ControlPlane{
@@ -77,9 +82,11 @@ func TestCreateControlPlane(t *testing.T) {
 					},
 				}
 
+				expectedRequest := cp.Spec.CreateControlPlaneRequest
+				expectedRequest.Labels = WithKubernetesMetadataLabels(cp, expectedRequest.Labels)
 				sdk.
 					EXPECT().
-					CreateControlPlane(ctx, cp.Spec.CreateControlPlaneRequest).
+					CreateControlPlane(ctx, expectedRequest).
 					Return(
 						nil,
 						&sdkkonnecterrs.BadRequestError{
@@ -280,7 +287,7 @@ func TestUpdateControlPlane(t *testing.T) {
 							Description: cp.Spec.Description,
 							AuthType:    (*sdkkonnectcomp.UpdateControlPlaneRequestAuthType)(cp.Spec.AuthType),
 							ProxyUrls:   cp.Spec.ProxyUrls,
-							Labels:      cp.Spec.Labels,
+							Labels:      WithKubernetesMetadataLabels(cp, cp.Spec.Labels),
 						},
 					).
 					Return(
@@ -333,7 +340,7 @@ func TestUpdateControlPlane(t *testing.T) {
 							Description: cp.Spec.Description,
 							AuthType:    (*sdkkonnectcomp.UpdateControlPlaneRequestAuthType)(cp.Spec.AuthType),
 							ProxyUrls:   cp.Spec.ProxyUrls,
-							Labels:      cp.Spec.Labels,
+							Labels:      WithKubernetesMetadataLabels(cp, cp.Spec.Labels),
 						},
 					).
 					Return(
@@ -386,7 +393,7 @@ func TestUpdateControlPlane(t *testing.T) {
 							Description: cp.Spec.Description,
 							AuthType:    (*sdkkonnectcomp.UpdateControlPlaneRequestAuthType)(cp.Spec.AuthType),
 							ProxyUrls:   cp.Spec.ProxyUrls,
-							Labels:      cp.Spec.Labels,
+							Labels:      WithKubernetesMetadataLabels(cp, cp.Spec.Labels),
 						},
 					).
 					Return(
@@ -397,9 +404,11 @@ func TestUpdateControlPlane(t *testing.T) {
 						},
 					)
 
+				expectedRequest := cp.Spec.CreateControlPlaneRequest
+				expectedRequest.Labels = WithKubernetesMetadataLabels(cp, expectedRequest.Labels)
 				sdk.
 					EXPECT().
-					CreateControlPlane(ctx, cp.Spec.CreateControlPlaneRequest).
+					CreateControlPlane(ctx, expectedRequest).
 					Return(
 						&sdkkonnectops.CreateControlPlaneResponse{
 							ControlPlane: &sdkkonnectcomp.ControlPlane{
@@ -441,4 +450,63 @@ func TestUpdateControlPlane(t *testing.T) {
 			assert.NoError(t, err)
 		})
 	}
+}
+
+func TestCreateAndUpdateControlPlane_KubernetesMetadataConsistency(t *testing.T) {
+	var (
+		ctx = context.Background()
+		cp  = &konnectv1alpha1.KonnectGatewayControlPlane{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       "KonnectGatewayControlPlane",
+				APIVersion: "konnect.konghq.com/v1alpha1",
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Name:       "cp-1",
+				Namespace:  "default",
+				UID:        k8stypes.UID(uuid.NewString()),
+				Generation: 2,
+			},
+			Spec: konnectv1alpha1.KonnectGatewayControlPlaneSpec{
+				CreateControlPlaneRequest: sdkkonnectcomp.CreateControlPlaneRequest{
+					Name: "cp-1",
+				},
+			},
+		}
+		sdk = NewMockControlPlaneSDK(t)
+	)
+
+	t.Log("Triggering CreateControlPlane with expected labels")
+	expectedLabels := map[string]string{
+		"k8s-name":       "cp-1",
+		"k8s-namespace":  "default",
+		"k8s-uid":        string(cp.GetUID()),
+		"k8s-kind":       "KonnectGatewayControlPlane",
+		"k8s-group":      "konnect.konghq.com",
+		"k8s-version":    "v1alpha1",
+		"k8s-generation": "2",
+	}
+	sdk.EXPECT().
+		CreateControlPlane(ctx, sdkkonnectcomp.CreateControlPlaneRequest{
+			Name:   "cp-1",
+			Labels: expectedLabels,
+		}).
+		Return(&sdkkonnectops.CreateControlPlaneResponse{
+			ControlPlane: &sdkkonnectcomp.ControlPlane{
+				ID: "12345",
+			},
+		}, nil)
+	require.NoError(t, createControlPlane(ctx, sdk, cp))
+
+	t.Log("Triggering UpdateControlPlane with expected labels")
+	sdk.EXPECT().
+		UpdateControlPlane(ctx, "12345", sdkkonnectcomp.UpdateControlPlaneRequest{
+			Name:   lo.ToPtr("cp-1"),
+			Labels: expectedLabels,
+		}).
+		Return(&sdkkonnectops.UpdateControlPlaneResponse{
+			ControlPlane: &sdkkonnectcomp.ControlPlane{
+				ID: "12345",
+			},
+		}, nil)
+	require.NoError(t, updateControlPlane(ctx, sdk, cp))
 }