+++ controller/konnect/ops/ops_kongconsumergroup.go
@@ -9,7 +9,6 @@ import (
 	sdkkonnectops "github.com/Kong/sdk-konnect-go/models/operations"
 	sdkkonnecterrs "github.com/Kong/sdk-konnect-go/models/sdkerrors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
 
@@ -17,7 +16,6 @@ import (
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
 
 	configurationv1beta1 "github.com/kong/kubernetes-configuration/api/configuration/v1beta1"
-	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
 	"github.com/kong/kubernetes-configuration/pkg/metadata"
 )
 
@@ -73,36 +71,16 @@ func createConsumerGroup(
 func updateConsumerGroup(
 	ctx context.Context,
 	sdk ConsumerGroupSDK,
-	cl client.Client,
 	group *configurationv1beta1.KongConsumerGroup,
 ) error {
-	if group.Spec.ControlPlaneRef == nil {
-		return fmt.Errorf("can't update %T without a ControlPlaneRef", group)
+	cpID := group.GetControlPlaneID()
+	if cpID == "" {
+		return fmt.Errorf("can't update %T %s without a Konnect ControlPlane ID", group, client.ObjectKeyFromObject(group))
 	}
 
-	// TODO(pmalek) handle other types of CP ref
-	// TODO(pmalek) handle cross namespace refs
-	nnCP := types.NamespacedName{
-		Namespace: group.Namespace,
-		Name:      group.Spec.ControlPlaneRef.KonnectNamespacedRef.Name,
-	}
-	var cp konnectv1alpha1.KonnectGatewayControlPlane
-	if err := cl.Get(ctx, nnCP, &cp); err != nil {
-		return fmt.Errorf("failed to get KonnectGatewayControlPlane %s: for %T %s: %w",
-			nnCP, group, client.ObjectKeyFromObject(group), err,
-		)
-	}
-
-	if cp.Status.ID == "" {
-		return fmt.Errorf(
-			"can't update %T when referenced KonnectGatewayControlPlane %s does not have the Konnect ID",
-			group, nnCP,
-		)
-	}
-
-	resp, err := sdk.UpsertConsumerGroup(ctx,
+	_, err := sdk.UpsertConsumerGroup(ctx,
 		sdkkonnectops.UpsertConsumerGroupRequest{
-			ControlPlaneID:  cp.Status.ID,
+			ControlPlaneID:  cpID,
 			ConsumerGroupID: group.GetKonnectStatus().GetKonnectID(),
 			ConsumerGroup:   kongConsumerGroupToSDKConsumerGroupInput(group),
 		},
@@ -125,8 +103,6 @@ func updateConsumerGroup(
 		return errWrapped
 	}
 
-	group.Status.Konnect.SetKonnectID(*resp.ConsumerGroup.ID)
-	group.Status.Konnect.SetControlPlaneID(cp.Status.ID)
 	k8sutils.SetCondition(
 		k8sutils.NewConditionWithGeneration(
 			conditions.KonnectEntityProgrammedConditionType,
@@ -180,7 +156,7 @@ func kongConsumerGroupToSDKConsumerGroupInput(
 	group *configurationv1beta1.KongConsumerGroup,
 ) sdkkonnectcomp.ConsumerGroupInput {
 	return sdkkonnectcomp.ConsumerGroupInput{
-		Tags: metadata.ExtractTags(group),
+		Tags: append(metadata.ExtractTags(group), GenerateKubernetesMetadataTags(group)...),
 		Name: group.Spec.Name,
 	}
 }