+++ controller/konnect/ops/ops_kongservice_test.go
@@ -7,10 +7,12 @@ import (
 	sdkkonnectcomp "github.com/Kong/sdk-konnect-go/models/components"
 	sdkkonnectops "github.com/Kong/sdk-konnect-go/models/operations"
 	sdkkonnecterrs "github.com/Kong/sdk-konnect-go/models/sdkerrors"
+	"github.com/google/uuid"
 	"github.com/samber/lo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8stypes "k8s.io/apimachinery/pkg/types"
 
 	"github.com/kong/gateway-operator/controller/konnect/conditions"
 	k8sutils "github.com/kong/gateway-operator/pkg/utils/kubernetes"
@@ -493,4 +495,48 @@ func TestUpdateKongService(t *testing.T) {
 			assert.NoError(t, err)
 		})
 	}
+}
+
+func TestCreateAndUpdateKongService_KubernetesMetadataConsistency(t *testing.T) {
+	svc := &configurationv1alpha1.KongService{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "KongService",
+			APIVersion: "configuration.konghq.com/v1alpha1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:       "svc-1",
+			Namespace:  "default",
+			UID:        k8stypes.UID(uuid.NewString()),
+			Generation: 2,
+			Annotations: map[string]string{
+				"konghq.com/tags": "tag1,tag2,duplicate-tag",
+			},
+		},
+		Status: configurationv1alpha1.KongServiceStatus{
+			Konnect: &konnectv1alpha1.KonnectEntityStatusWithControlPlaneRef{
+				ControlPlaneID: uuid.NewString(),
+			},
+		},
+		Spec: configurationv1alpha1.KongServiceSpec{
+			KongServiceAPISpec: configurationv1alpha1.KongServiceAPISpec{
+				Tags: []string{"tag3", "tag4", "duplicate-tag"},
+			},
+		},
+	}
+	output := kongServiceToSDKServiceInput(svc)
+	expectedTags := []string{
+		"k8s-kind:KongService",
+		"k8s-name:svc-1",
+		"k8s-uid:" + string(svc.GetUID()),
+		"k8s-version:v1alpha1",
+		"k8s-group:configuration.konghq.com",
+		"k8s-namespace:default",
+		"k8s-generation:2",
+		"tag1",
+		"tag2",
+		"tag3",
+		"tag4",
+		"duplicate-tag",
+	}
+	require.ElementsMatch(t, expectedTags, output.Tags)
 }